// src/app/shared/enums/security-permission-type.enum.ts
export enum SecurityPermissionType {
  All = 1,
  FullAccess = 2,
  CostingModule = 3,
  CostingRead = 4,
  CostingWrite = 5,
  DFModule = 6,
  DFRead = 7,
  DFWrite = 8,
  ProjectModule = 9,
  ProjectRead = 10,
  ProjectWrite = 11,
  AIModule = 12,
  AIRead = 13,
  AIWrite = 14,
  SupplierModule = 15,
  SupplierRead = 16,
  SupplierWrite = 17,
  ReportsModule = 18,
  ReportsRead = 19,
  ReportsWrite = 20,
  UserModule = 21,
  UserRead = 22,
  UserWrite = 23,
  CustomerModule = 24,
  CustomerRead = 25,
  CustomerWrite = 26,
  AnalyticsModule = 27,
  AnalyticsRead = 28,
  AnalyticsWrite = 29,
}

//Adduser, save --user
//save -- customer